Move test_feather and test_json to directories
